feat: avatarnumber db에 저장
