+++ src/app/layout.tsx
@@ -1,14 +1,16 @@
-'use client';
 import './globals.css';
 import styles from './styles/global.module.scss';
-import { SessionProvider } from 'next-auth/react';
-import { Session } from 'next-auth';
 
-export default function RootLayout({ children, session }: { children: React.ReactNode; session: Session }) {
+import { getServerSession, Session } from 'next-auth';
+import { authOptions } from './utils/authOptions';
+import SessionProviderWrapper from './Providers';
+
+export default async function RootLayout({ children }: { children: React.ReactNode }) {
+  const session: Session | null = await getServerSession(authOptions);
   return (
     <html lang="en">
-      <body className={` ${styles.container}`}>
-        <SessionProvider session={session}>{children}</SessionProvider>
+      <body className={styles.container}>
+        <SessionProviderWrapper session={session}>{children}</SessionProviderWrapper>
       </body>
     </html>
   );
