+++ src/app/page.tsx
@@ -17,6 +17,7 @@ export default function Home() {
     <div className={styles.contentContainer}>
       <Image className={styles.logo} src="/logo.svg" alt="LOMO logo" width={263} height={77} priority />
       <div className={styles.textBox}>
+        <div></div>
         <div>&quot;내 몸을 알고 나를 알다&quot;</div>
         <div>체형분석을 통해 당신만의 특별함을 찾아보세요</div>
       </div>
