+++ src/app/services/client/updateAvatarNumber.ts
@@ -1,0 +1,23 @@
+export async function updateAvatarNumber(userEmail: string, avatarNumber: number) {
+  try {
+    const res = await fetch('/api/user', {
+      method: 'PUT',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        userEmail: userEmail,
+        avatarNumber: avatarNumber,
+      }),
+    });
+
+    if (!res.ok) {
+      throw new Error(`HTTP error! Status: ${res.status}`);
+    }
+
+    const data = await res.json();
+    return data;
+  } catch (error) {
+    console.error('Failed to update avatar number:', error);
+  }
+}
